examples/api_buyer/requirements.txt
agentstr==0.1.13
python-dotenv
fastapi
uvicorn
cassandra-driver
cassio